+++ tests/phpunit/Unit/Components/MessageTest.php
@@ -1,0 +1,44 @@
+<?php
+/**
+ * This file is part of the MediaWiki skin Chameleon.
+ *
+ * @copyright 2021, Morne Alberts
+ * @license   GPL-3.0-or-later
+ *
+ * The Chameleon skin is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by the Free
+ * Software Foundation, either version 3 of the License, or (at your option) any
+ * later version.
+ *
+ * The Chameleon skin is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+ * details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @file
+ * @ingroup Skins
+ */
+
+namespace Skins\Chameleon\Tests\Unit\Components;
+
+/**
+ * @coversDefaultClass \Skins\Chameleon\Components\Message
+ * @covers ::<private>
+ * @covers ::<protected>
+ *
+ * @group   skins-chameleon
+ * @group   mediawiki-databaseless
+ *
+ * @author Morne Alberts
+ * @since 3.4
+ * @ingroup Skins
+ * @ingroup Test
+ */
+class MessageTest extends GenericComponentTestCase {
+
+	protected $classUnderTest = '\Skins\Chameleon\Components\Message';
+
+}