+++ docs/release-notes.md
@@ -6,7 +6,7 @@ Under development.
 
 * Added `Message` component to display system messages (thanks @malberts)
 * Added `PersonalTools` component user name display support via the new `showUserName` attribute (thanks @malberts)
-* Fixed `Toolbox` component PHP notice (thanks @WouterRademaker)
+* Fixed `Toolbox` component PHP notice (thanks @WouterRademaker and @malberts)
 * Translation updates for system messages (thanks @translatewiki and its translator community)
 
 ### Chameleon 3.3.0