+++ docs/release-notes.md
@@ -10,7 +10,12 @@ Under development.
 
 ### Chameleon 3.4.0
 
-Under development.
+Released on August 7, 2021.
+
+* Added `Message` component to display system messages (thanks @malberts)
+* Added `PersonalTools` component user name display support via the new `showUserName` attribute (thanks @malberts)
+* Fixed `Toolbox` component PHP notice (thanks @WouterRademaker and @malberts)
+* Translation updates for system messages (thanks @translatewiki and its translator community)
 
 ### Chameleon 3.3.0